order-service/requirements.txt
fastapi
httpx
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
